@@ -5,5 +5,5 @@ pythonpath = [
 
 [tool.black]
 line-length = 110
-target-version = ['py37']
+target-version = ['py310']
 include = '\.pyi?$'
