@@ -1,18 +1,20 @@
 import json
+from typing import List
+
 import yaml
-from typing import Dict, List, Optional, Union
 
 from src.ai_tools.models.file_spec import FileSpec
-from src.processors.api_processor import APIProcessor
-from src.services.file_service import FileService
 from src.configuration.config import Config
+from src.processors.api_processor import APIProcessor
+from src.models import APIModel, APIPath, APIVerb, GeneratedModel, ModelInfo, APIDefinition
+from src.services.file_service import FileService
 from .swagger import (
     APIDefinitionMerger,
     APIDefinitionSplitter,
+    FileLoader,
     APIDefinitionLoader,
 )
 from ..utils.logger import Logger
-import re
 
 
 class SwaggerProcessor(APIProcessor):
@@ -25,7 +27,7 @@ class SwaggerProcessor(APIProcessor):
         merger: APIDefinitionMerger,
         file_service: FileService,
         config: Config,
-        apiDefinitionLoader: APIDefinitionLoader = None,
+        api_definition_loader: APIDefinitionLoader = None,
     ):
         """
         Initialize the SwaggerProcessor.
@@ -34,17 +36,17 @@ class SwaggerProcessor(APIProcessor):
             file_loader (FileLoader): Service to load API definition files.
             splitter (APIDefinitionSplitter): Service to split API definitions.
             merger (APIDefinitionMerger): Service to merge API definitions.
-            apiDefinitionLoader (APIDefinitionLoader): Service to load API definition from URL or file.
+            api_definition_loader (APIDefinitionLoader): Service to load API definition from URL or file.
         """
         self.config = config
         self.file_service = file_service
         self.file_loader = file_loader
         self.splitter = splitter
         self.merger = merger
-        self.apiDefinitionLoader = apiDefinitionLoader or APIDefinitionLoader()
+        self.api_definition_loader = api_definition_loader or APIDefinitionLoader()
         self.logger = Logger.get_logger(__name__)
 
-    def process_api_definition(self, api_definition: str) -> List[Dict[str, Union[str, Dict]]]:
+    def process_api_definition(self, api_definition: str) -> APIDefinition:
         """
         Processes an API definition by loading, splitting, and merging its components.
 
@@ -52,30 +54,38 @@ class SwaggerProcessor(APIProcessor):
             api_definition (str): URL or path to the API definition.
 
         Returns:
-            List of merged API definitions.
+            APIDefinition containing the processed API definitions.
         """
         try:
             self.logger.info("Starting API processing")
-
-            raw_definition = self.apiDefinitionLoader.load(api_definition)
+            raw_definition = self.api_definition_loader.load(api_definition)
             split_definitions = self.splitter.split(raw_definition)
             merged_definitions = self.merger.merge(split_definitions)
 
+            result = APIDefinition(endpoints=self.config.endpoints)
             for definition in merged_definitions:
-                self.logger.debug(f"\nType: {definition['type']}")
-                self.logger.debug(f"Path: {definition['path']}")
-                self.logger.debug(f"Verb: {definition['verb']}")
+                if definition.type == "path":
+                    result.add_definition(APIPath(path=definition.path, yaml=definition.yaml))
+                elif definition.type == "verb":
+                    result.add_definition(
+                        APIVerb(
+                            verb=definition.verb,
+                            path=definition.path,
+                            root_path=APIVerb.get_root_path(definition.path),
+                            yaml=definition.yaml,
+                        )
+                    )
 
             self.logger.info("Successfully processed API definition.")
-            return merged_definitions
+            return result
         except Exception as e:
             self.logger.error(f"Error processing API definition: {e}")
             raise
 
-    def extract_env_vars(self, api_definitions: List[Dict[str, Union[str, Dict]]]) -> None:
+    def extract_env_vars(self, api_definition: APIDefinition) -> list[str]:
         self.logger.info("\nGenerating .env file...")
 
-        api_definition_str = api_definitions[0]["yaml"]
+        api_definition_str = api_definition.definitions[0].yaml
         try:
             api_spec = json.loads(api_definition_str)
         except json.JSONDecodeError:
@@ -94,8 +104,10 @@ class SwaggerProcessor(APIProcessor):
         self.file_service.create_files(self.config.destination_folder, [file_spec])
 
         self.logger.info(f"Generated .env file with BASEURL={base_url}")
+        return []
 
-    def _extract_base_url(self, api_spec):
+    @staticmethod
+    def _extract_base_url(api_spec):
         """Extract base URL from OpenAPI specification"""
         if "openapi" in api_spec and api_spec["openapi"].startswith("3."):
             if "servers" in api_spec and api_spec["servers"] and "url" in api_spec["servers"][0]:
@@ -112,90 +124,65 @@ class SwaggerProcessor(APIProcessor):
 
         return None
 
-    def get_api_paths(
-        self, api_definition: Union[str, Dict], endpoints: Optional[List[str]] = None
-    ) -> List[Dict[str, Union[str, Dict]]]:
-        result = []
+    def get_api_paths(self, api_definition: APIDefinition) -> List[APIPath]:
+        """Get all path definitions that should be processed."""
+        return api_definition.get_filtered_paths()
 
-        for definition in api_definition:
-            if not self._should_process_endpoint(definition["path"], endpoints):
-                continue
-            if definition["type"] == "path":
-                result.append(definition)
+    def get_api_path_name(self, api_path: APIPath) -> str:
+        return api_path.path
 
-        return result
+    def get_api_verbs(self, api_definition: APIDefinition) -> List[APIVerb]:
+        """Get all verb definitions that should be processed."""
+        return api_definition.get_filtered_verbs()
 
-    def _should_process_endpoint(self, path: str, endpoints: List[str]) -> bool:
-        """Check if an endpoint should be processed based on configuration"""
-        if endpoints is None:
-            return True
+    def get_api_verb_path(self, api_verb: APIVerb) -> str:
+        return api_verb.path
 
-        return any(path.startswith(endpoint) for endpoint in endpoints)
+    def get_api_verb_rootpath(self, api_verb: APIVerb) -> str:
+        return api_verb.root_path
 
-    def get_api_path_name(self, api_path: Dict[str, Union[str, Dict]]) -> str:
-        return api_path["path"]
+    def get_api_verb_name(self, api_verb: APIVerb) -> str:
+        return api_verb.verb
 
-    def get_api_verbs(
-        self, api_definition: Dict[str, str], endpoints: Optional[List[str]] = None
-    ) -> List[Dict[str, Union[str, Dict]]]:
-        result = []
+    def get_relevant_models(self, all_models: List[ModelInfo], api_verb: APIVerb) -> List[GeneratedModel]:
+        """Get models relevant to the API verb."""
+        try:
+            self.logger.info(f"Getting relevant models for {api_verb.path} {api_verb.verb}")
 
-        for definition in api_definition:
-            if not self._should_process_endpoint(definition["path"], endpoints):
-                continue
-            if definition["type"] == "verb":
-                result.append(definition)
+            # Find the model info for this path
+            path_model_info = next(
+                (info for info in all_models if info.path == api_verb.path), ModelInfo(path=api_verb.path)
+            )
 
-        return result
+            # Get models by path
+            relevant_models = path_model_info.get_models_by_path(api_verb.path)
 
-    def get_api_verb_path(self, api_verb_definition: Dict[str, Union[str, Dict]]) -> str:
-        return api_verb_definition["path"]
+            # If no models found by path, try to find by summary
+            if not relevant_models:
+                relevant_models = path_model_info.get_models_by_summary(api_verb.verb.lower())
 
-    def get_api_verb_rootpath(self, api_verb_definition: Dict[str, Union[str, Dict]]) -> str:
-        return self._get_root_path(api_verb_definition["path"])
+            self.logger.info(
+                f"Found {len(relevant_models)} relevant models for {api_verb.path} {api_verb.verb}"
+            )
+            return relevant_models
 
-    def get_api_verb_name(self, api_verb: Dict[str, Union[str, Dict]]) -> str:
-        return api_verb["verb"]
-
-    def _get_root_path(self, path: str) -> str:
-        match = re.match(r"(/[^/?]+)", path)
-        if match:
-            return match.group(1)
-        return path
-
-    def get_relevant_models(
-        self, all_models: List[Dict[str, Union[str, Dict]]], api_verb: Union[str, Dict]
-    ) -> List[Dict[str, Union[str, Dict]]]:
-        result = []
-
-        for model in all_models:
-            if api_verb["path"] == model["path"] or str(api_verb["path"]).startswith(model["path"] + "/"):
-                result.append(model["models"])
-
-        return result
+        except Exception as e:
+            self.logger.error(f"Error getting relevant models: {str(e)}")
+            return []
 
     def get_other_models(
         self,
-        all_models: List[Dict[str, Union[str, Dict]]],
-        api_verb: Dict[str, Union[str, Dict]],
-    ) -> List[Dict[str, str]]:
+        all_models: List[ModelInfo],
+        api_verb: APIVerb,
+    ) -> List[APIModel]:
         result = []
-
         for model in all_models:
-            if not (
-                api_verb["path"] == model["path"] or str(api_verb["path"]).startswith(model["path"] + "/")
-            ):
-                result.append(
-                    {
-                        "path": model["path"],
-                        "files": model["files"],
-                    }
-                )
-
+            if not (api_verb.path == model.path or str(api_verb.path).startswith(model.path + "/")):
+                result.append(APIModel(path=model.path, files=model.files))
         return result
 
-    def get_api_verb_content(self, api_verb: Dict[str, Union[str, Dict]]) -> str:
-        return api_verb["yaml"]
+    def get_api_verb_content(self, api_verb: APIVerb) -> str:
+        return api_verb.yaml
 
-    def get_api_path_content(self, api_path: Dict[str, Union[str, Dict]]) -> str:
-        return api_path["yaml"]
+    def get_api_path_content(self, api_path: APIPath) -> str:
+        return api_path.yaml
