@@ -4,6 +4,4 @@ from pydantic import BaseModel, Field
 
 
 class FileReadingInput(BaseModel):
-    files: List[str] = Field(
-        description="A list of file paths to be read."
-    )
+    files: List[str] = Field(description="A list of file paths to be read.")
