@@ -1,387 +1,122 @@
 import json
 import os
-import re
-import copy
-from typing import Dict, List, Optional, Tuple, Union
+from typing import List
 
 from src.ai_tools.models.file_spec import FileSpec
 from src.processors.api_processor import APIProcessor
+from src.models import APIModel, APIPath, APIVerb, GeneratedModel, ModelInfo, APIDefinition
+from src.processors.postman.models import RequestData
+from src.processors.postman.postman_utils import PostmanUtils
 from src.services.file_service import FileService
 from src.utils.logger import Logger
 
 
 class PostmanProcessor(APIProcessor):
-    """Processes V2 Postman collection .json"""
-
-    service_dict = {}
-    numeric_only = r"^\d+$"
+    """Processes Postman API definitions."""
 
     def __init__(self, file_service: FileService):
         self.file_service = file_service
         self.logger = Logger.get_logger(__name__)
 
-    def process_api_definition(self, json_file_path: str) -> List[Dict[str, Union[str, Dict]]]:
-        with open(json_file_path, encoding="utf-8") as postman_json_export:
-            data = json.load(postman_json_export)
-            return self.extract_requests(data)
+    def process_api_definition(self, json_file_path: str) -> APIDefinition:
+        with open(json_file_path, encoding="utf-8") as f:
+            data = json.load(f)
+        return APIDefinition(PostmanUtils.extract_requests(data))
 
-    def extract_env_vars(self, extracted_requests: List[Dict[str, Union[str, Dict]]]) -> List[str]:
-        env_vars = set()
-        for request in extracted_requests:
-            path = request.get("path", "")
-            if path.startswith("{{"):
-                match = re.match(r"\{\{(.*?)\}\}", path)
-                if match:
-                    env_vars.add(match.group(1))
-                    break
-        return list(env_vars)
+    def extract_env_vars(self, api_definition: APIDefinition) -> List[str]:
+        return PostmanUtils.extract_env_vars(api_definition)
 
-    def get_api_paths(
-        self, api_definition: Dict[str, Union[str, Dict]], endpoints=Optional[List[str]]
-    ) -> List[Tuple]:
-        all_paths_no_query_params = self.get_all_distinct_paths_no_query_params(api_definition)
-        verb_path_pairs = self._extract_verb_path_info(api_definition)
-        paths_grouped_by_service = self._group_paths_by_service(all_paths_no_query_params)
-        self.service_dict = self.map_verb_path_pairs_to_services(verb_path_pairs, paths_grouped_by_service)
+    def get_api_paths(self, api_definition: APIDefinition) -> List[APIPath]:
+        # Build VerbInfo list
+        verbs = PostmanUtils.extract_verb_path_info(api_definition.definitions)
+        # Build ServiceVerbs
+        paths = [r.path.split("?", 1)[0] for r in api_definition.definitions]
+        sv = PostmanUtils.map_verb_path_pairs_to_services(verbs, paths)
 
-        return copy.deepcopy(self.service_dict).items()
-
-    def get_api_path_name(self, api_path: Tuple) -> str:
-        return api_path[0]
-
-    def get_relevant_models(
-        self,
-        all_models: List[Dict[str, Union[str, Dict]]],
-        api_verb: Dict[str, Union[str, Dict]],
-    ) -> List[str]:
         result = []
-
-        for model in all_models:
-            if api_verb["service"] == model["path"]:
-                result = [file["fileContent"] for file in model["models"]]
-
+        for service in sv:
+            result.append(APIPath(path=service.service, yaml=json.dumps({"path": service.service})))
         return result
+
+    def get_api_path_name(self, api_path: APIPath) -> str:
+        return api_path.path
+
+    def get_relevant_models(self, all_models: List[ModelInfo], api_verb: APIVerb) -> List[GeneratedModel]:
+        """Get models relevant to the API verb."""
+        try:
+            self.logger.info(f"Getting relevant models for {api_verb.path} {api_verb.verb}")
+
+            # Find the model info for this path
+            path_model_info = next(
+                (info for info in all_models if info.path == api_verb.path), ModelInfo(path=api_verb.path)
+            )
+
+            # Get models by path
+            relevant_models = path_model_info.get_models_by_path(api_verb.path)
+
+            # If no models found by path, try to find by summary
+            if not relevant_models:
+                relevant_models = path_model_info.get_models_by_summary(api_verb.verb.lower())
+
+            self.logger.info(
+                f"Found {len(relevant_models)} relevant models for {api_verb.path} {api_verb.verb}"
+            )
+            return relevant_models
+
+        except Exception as e:
+            self.logger.error(f"Error getting relevant models: {str(e)}")
+            return []
 
     def get_other_models(
-        self, all_models: List[Dict[str, Union[str, Dict]]], api_verb: Dict[str, Union[str, Dict]]
-    ) -> List[Dict[str, List[str]]]:
+        self,
+        all_models: List[ModelInfo],
+        verb: APIVerb,
+    ) -> List[APIModel]:
         result = []
-
         for model in all_models:
-            if not api_verb["service"] == model["path"]:
-                result.append(
-                    {
-                        "files": model["files"],
-                    }
-                )
-
-    def _group_paths_by_service(self, all_distinct_paths_no_query_params: List[str]) -> Dict[str, List[str]]:
-        grouped_paths_dict = {}
-        for path in all_distinct_paths_no_query_params:
-            root_path = path.split("/")[1]
-            if root_path not in grouped_paths_dict:
-                grouped_paths_dict[root_path] = []
-            grouped_paths_dict[root_path].append(path)
-
-        return grouped_paths_dict
-
-    def get_api_verb_path(self, api_verb_definition: Dict[str, Union[str, Dict]]) -> str:
-        return api_verb_definition["path"]
-
-    def get_api_verb_name(self, api_verb_definition: Dict[str, Union[str, Dict]]) -> str:
-        return api_verb_definition["verb"]
-
-    def get_api_verb_rootpath(self, api_verb_definition: Dict[str, Union[str, Dict]]) -> str:
-        return api_verb_definition["service"]
-
-    def _contains_same_items(self, array1: List, array2: List) -> bool:
-        """
-        Check if two arrays contain the same items
-        """
-        return sorted(array1) == sorted(array2)
-
-    def _get_all_paths_from_service_dict(self, service_dict: Dict[str, List[str]]) -> List[str]:
-        all_paths = []
-        for paths in service_dict.values():
-            all_paths.extend(paths)
-        return all_paths
-
-    def get_api_verbs(
-        self, api_definition: Dict[str, str], endpoints=None
-    ) -> List[Dict[str, Union[str, Dict]]]:
-        return self._add_service_name_to_verb_chunks(api_definition, self.service_dict)
-
-    def get_api_verb_content(self, api_verb: Dict[str, Union[str, Dict]]) -> Dict[str, Union[str, Dict]]:
-        return api_verb
-
-    def get_api_path_content(self, api_path: Dict[str, Union[str, Dict]]) -> Dict[str, Union[str, Dict]]:
-        return api_path
-
-    def extract_requests(self, data, path=""):
-        results = []
-
-        def process_item(item, current_path):
-            if isinstance(item, dict):
-                if "item" in item and isinstance(item["item"], list):
-                    new_path = (
-                        f'{current_path}/{self._to_camel_case(item["name"])}'
-                        if "name" in item
-                        else current_path
-                    )
-                    for sub_item in item["item"]:
-                        process_item(sub_item, new_path)
-                elif self._item_is_a_test_case(item):
-                    result = self._extract_request_data(item, current_path)
-                    if result["name"] not in {r["name"] for r in results}:
-                        results.append(result)
-                else:
-                    for sub_key, sub_value in item.items():
-                        process_item(sub_value, current_path)
-            elif isinstance(item, list):
-                for sub_item in item:
-                    process_item(sub_item, current_path)
-
-        process_item(data, path)
-        return results
-
-    def map_verb_path_pairs_to_services(
-        self,
-        verb_path_pairs: Dict[str, Union[List, Dict, str]],
-        no_query_params_routes_grouped_by_service: Dict[str, List[str]],
-    ) -> Dict[str, Union[List, Dict, str]]:
-        verb_chunks_with_query_params = self._extract_verb_path_info(verb_path_pairs)
-
-        verb_path_pairs_and_services = {}
-
-        for verb_path_pair in verb_chunks_with_query_params:
-
-            for service, routes in no_query_params_routes_grouped_by_service.items():
-
-                if verb_path_pair["path"] in routes:
-
-                    if service not in verb_path_pairs_and_services:
-                        verb_path_pairs_and_services[service] = []
-
-                    verb_path_pairs_and_services[service].append(
-                        {
-                            "verb": verb_path_pair["verb"],
-                            "path": verb_path_pair["path"],
-                            "query_params": verb_path_pair["query_params"],
-                            "body": verb_path_pair["body"],
-                        }
-                    )
-
-        return verb_path_pairs_and_services
-
-    def _add_service_name_to_verb_chunks(
-        self,
-        verb_chunks: List[Dict[str, Union[List, Dict, str]]],
-        all_services_dict: Dict[str, List[str]],
-    ) -> List[Dict[str, Union[List, Dict, str]]]:
-        verb_chunks_tagged_with_service = copy.deepcopy(verb_chunks)
-
-        for verb_chunk in verb_chunks_tagged_with_service:
-
-            for service, verbs in all_services_dict.items():
-
-                routes_in_service = [verb["path"] for verb in verbs]
-
-                verb_chunk_path_no_query_params = verb_chunk["path"].split("?")[0]
-
-                if verb_chunk_path_no_query_params in routes_in_service:
-                    verb_chunk["service"] = service
-                    break
-
-        return verb_chunks_tagged_with_service
-
-    def get_all_distinct_paths_no_query_params(
-        self, extracted_verb_chunks: List[Dict[str, Union[List, Dict, str]]]
-    ) -> List[str]:
-        distinct_paths_no_query_params = list(
-            set([item["path"].split("?")[0] for item in extracted_verb_chunks])
-        )
-
-        return distinct_paths_no_query_params
-
-    def _extract_verb_path_info(
-        self, extracted_requests: List[Dict[str, Union[List, Dict, str]]]
-    ) -> List[Dict[str, Union[Dict, str]]]:
-        distinct_paths_no_query_params = self.get_all_distinct_paths_no_query_params(extracted_requests)
-        result = []
-
-        for path_no_query_params in distinct_paths_no_query_params:
-            verbs = []
-            matching_full_paths = []
-
-            for item in extracted_requests:
-                if item["path"].startswith(path_no_query_params):
-                    matching_full_paths.append(item)
-
-                    if item["verb"] not in verbs:
-                        verbs.append(item["verb"])
-
-            for verb in verbs:
-                all_query_params_on_verb_path = {}
-                all_body_attributes_on_verb_path = {}
-
-                for path_item in matching_full_paths:
-                    if path_item["verb"] == verb:
-
-                        if path_item["body"] is not None:
-                            self._accumulate_request_body_attributes(
-                                all_body_attributes_on_verb_path, path_item["body"]
-                            )
-
-                        path_sliced_on_query_param_start = path_item["path"].split("?")
-
-                        if len(path_sliced_on_query_param_start) > 1:
-                            all_query_params = path_sliced_on_query_param_start[1].split("&")
-                            if len(all_query_params) > 0:
-                                self._accumulate_query_params(all_query_params_on_verb_path, all_query_params)
-
-                result.append(
-                    {
-                        "verb": verb,
-                        "root_path": self._get_root_path(path_no_query_params),
-                        "path": path_no_query_params,
-                        "query_params": all_query_params_on_verb_path,
-                        "body": all_body_attributes_on_verb_path,
-                    }
-                )
-
+            if model.path != verb.root_path:
+                result.append(APIModel(path=model.path, files=model.files))
         return result
 
-    def _accumulate_query_params(self, all_query_params: Dict, current_request_query_params: str):
-        for param in current_request_query_params:
+    def get_api_verb_path(self, verb: APIVerb) -> str:
+        return verb.path
 
-            param_array = param.split("=")
-            param_name = param_array[0]
+    def get_api_verb_rootpath(self, verb: APIVerb) -> str:
+        return verb.root_path
 
-            if (param_name) and (param_name not in all_query_params):
-                if len(param_array) > 1:
-                    if re.match(PostmanProcessor.numeric_only, param_array[1]):
-                        all_query_params[param_name] = "number"
-                    else:
-                        all_query_params[param_name] = "string"
-            elif all_query_params[param_name] == "number" and len(param_array) > 1:
-                if not re.match(PostmanProcessor.numeric_only, param_array[1]):
-                    all_query_params[param_name] = "string"
+    def get_api_verb_name(self, verb: APIVerb) -> str:
+        return verb.verb
 
-    def _accumulate_request_body_attributes(self, all_body_attributes: Dict, current_request_body: Dict):
-        for key, value in current_request_body.items():
-            if key not in all_body_attributes:
-                if isinstance(value, str) and re.match(PostmanProcessor.numeric_only, value):
-                    all_body_attributes[key] = "number"
-                elif isinstance(value, str):
-                    all_body_attributes[key] = "string"
-                elif isinstance(value, dict):
-                    all_body_attributes[f"{key}Object"] = self._map_object_attributes(value)
-                elif isinstance(value, list):
-                    all_body_attributes[f"{key}Object"] = "array"
-            elif isinstance(value, str) and not re.match(PostmanProcessor.numeric_only, value):
-                all_body_attributes[key] = "string"
+    def get_api_verbs(self, api_definition: APIDefinition) -> List[APIVerb]:
+        return api_definition.get_verbs()
 
-    def _to_camel_case(self, s: str) -> str:
-        words = re.split(r"[^a-zA-Z0-9]", s)
-        words = [word for word in words if word]
-        if not words:
-            return ""
-        return words[0].lower() + "".join(word.capitalize() for word in words[1:])
+    def get_api_verb_content(self, verb: APIVerb) -> str:
+        return verb.yaml
 
-    def _extract_request_data(
-        self, data: Dict[str, Union[str, List, Dict]], path: str
-    ) -> List[Dict[str, Union[str, Dict, List]]]:
-        result = {
-            "file_path": "",
-            "path": "",
-            "verb": "",
-            "body": {},
-            "prerequest": [],
-            "script": [],
-            "name": "",
-        }
+    def get_api_path_content(self, svc: APIPath) -> str:
+        return svc.yaml
 
-        if "request" in data and isinstance(data["request"], dict):
-
-            result["verb"] = data["request"].get("method", "")
-
-            if "url" in data["request"]:
-                url_value = data["request"].get("url", "")
-                raw_url_value = url_value["raw"] if isinstance(url_value, dict) else url_value
-                result["path"] = raw_url_value
-
-            if "body" in data["request"]:
-                try:
-                    raw_body = data["request"]["body"].get("raw", "")
-                    clean_json_body = json.loads(raw_body.replace("\r", "").replace("\n", ""))
-                except (KeyError, json.JSONDecodeError):
-                    clean_json_body = None
-                result["body"] = clean_json_body
-
-        if "event" in data and isinstance(data["event"], list):
-            for script in data["event"]:
-                if script.get("listen") == "test":
-                    result["script"] = script.get("script", {}).get("exec", [])
-                elif script.get("listen") == "prerequest":
-                    result["prerequest"] = script.get("script", {}).get("exec", [])
-
-        result["name"] = self._to_camel_case(data["name"])
-        result["file_path"] = f"src/tests{path}/{result['name']}"
-
-        return result
-
-    def _item_is_a_test_case(self, data: str) -> bool:
-        return "request" in data or (("event" in data) and ("request" in data))
-
-    def _get_root_path(self, full_path: str) -> str:
-        match = re.search(r"(?<!/)/([^/]+)/", full_path)
-        if match:
-            return match.group(1)
-        else:
-            return ""
-
-    def _map_object_attributes(self, obj: Dict) -> Dict:
-        mapped_attributes = {}
-
-        for key, value in obj.items():
-            if isinstance(value, str) and re.match(PostmanProcessor.numeric_only, value):
-                mapped_attributes[key] = "number"
-            elif isinstance(value, str):
-                mapped_attributes[key] = "string"
-            elif isinstance(value, dict):
-                mapped_attributes[f"{key}Object"] = self._map_object_attributes(value)
-            elif isinstance(value, list):
-                mapped_attributes[f"{key}Object"] = "array"
-
-        return mapped_attributes
-
-    def update_framework_for_postman(self, destination_folder, extracted_requests):
-        self._create_run_order_file(destination_folder, extracted_requests)
+    def update_framework_for_postman(self, destination_folder: str, api_definition: APIDefinition):
+        self._create_run_order_file(destination_folder, api_definition)
         self._update_package_dot_json(destination_folder)
 
     def _update_package_dot_json(self, destination_folder: str):
-        package_json_path = os.path.join(destination_folder, "package.json")
-
+        pkg = os.path.join(destination_folder, "package.json")
         try:
-            with open(package_json_path, "r") as file:
-                package_json = json.load(file)
-                package_json["scripts"]["test"] = "mocha runTestsInOrder.js --timeout 10000"
-
-            with open(package_json_path, "w") as file:
-                json.dump(package_json, file, indent=2)
-
-            self.logger.info(f"Updated package.json at {package_json_path}")
+            with open(pkg, "r") as f:
+                data = json.load(f)
+            data.setdefault("scripts", {})["test"] = "mocha runTestsInOrder.js --timeout 10000"
+            with open(pkg, "w") as f:
+                json.dump(data, f, indent=2)
+            self.logger.info(f"Updated package.json at {pkg}")
         except Exception as e:
             self.logger.error(f"Failed to update package.json: {e}")
 
-    def _create_run_order_file(self, destination_folder: str, extracted_requests: List[str]):
-        file_content = ["// This file runs the tests in order"]
-
-        for request in extracted_requests:
-            filepath = f'./{request["file_path"]}.spec.ts'
-            file_content.append(f'import "{filepath}";')
-
-        run_tests_file_spec = FileSpec(path="runTestsInOrder.js", fileContent="\n".join(file_content))
-
-        self.file_service.create_files(destination_folder, [run_tests_file_spec])
-        self.logger.info(f"Created runTestsInOrder.js file at {destination_folder}")
+    def _create_run_order_file(self, destination_folder: str, api_definition: APIDefinition):
+        lines = ["// This file runs the tests in order"]
+        for definition in api_definition.definitions:
+            if isinstance(definition, APIVerb):
+                lines.append(f'import "./{definition.path}.spec.ts";')
+        file_spec = FileSpec(path="runTestsInOrder.js", fileContent="\n".join(lines))
+        self.file_service.create_files(destination_folder, [file_spec])
+        self.logger.info(f"Created runTestsInOrder.js at {destination_folder}")
