@@ -1,7 +1,7 @@
+from .api_definition_loader import APIDefinitionLoader
 from .api_definition_merger import APIDefinitionMerger
 from .api_definition_splitter import APIDefinitionSplitter
 from .file_handler import FileLoader
-from .api_definition_loader import APIDefinitionLoader
 
 __all__ = [
     "APIDefinitionMerger",
