@@ -1,8 +1,10 @@
 import copy
-from typing import Optional, Dict, List
+from typing import Dict, List
+import re
 
 import yaml
 
+from src.models import APIPath, APIVerb, APIDef
 from src.utils.logger import Logger
 
 
@@ -12,45 +14,38 @@ class APIDefinitionSplitter:
     def __init__(self):
         self.logger = Logger.get_logger(__name__)
 
-    def split(self, api_definition: Dict) -> List[Dict]:
+    def split(self, api_definition: Dict) -> List[APIDef]:
         """Splits the API definition into smaller, manageable parts."""
         self.logger.info("Splitting API definition into components...")
         api_definition_list = []
 
         base_copy = copy.deepcopy(api_definition)
         del base_copy["paths"]
-        api_definition_list.append(self._create_entry("base", None, None, base_copy))
+        api_definition_list.append(APIPath(path="/", yaml=yaml.dump(base_copy, sort_keys=False)))
 
         for path, path_data in api_definition.get("paths", {}).items():
-            normalized_path = self._normalize_path(path)
+            normalized_path = APIPath.normalize_path(path)
 
             path_copy = copy.deepcopy(api_definition)
             path_copy["paths"] = {path: path_data}
             api_definition_list.append(
-                self._create_entry("path", normalized_path, None, path_copy)
+                APIPath(path=normalized_path, yaml=yaml.dump(path_copy, sort_keys=False))
             )
 
             for verb, verb_data in path_data.items():
                 verb_copy = copy.deepcopy(path_copy)
                 verb_copy["paths"][path] = {verb: verb_data}
                 api_definition_list.append(
-                    self._create_entry("verb", normalized_path, verb.upper(), verb_copy)
+                    APIVerb(
+                        verb=verb.upper(),
+                        path=normalized_path,
+                        root_path=APIVerb.get_root_path(normalized_path),
+                        yaml=yaml.dump(verb_copy, sort_keys=False),
+                    )
                 )
 
         self.logger.info("Successfully split API definition.")
         return api_definition_list
-
-    @staticmethod
-    def _create_entry(
-        entry_type: str, path: Optional[str], verb: Optional[str], yaml_content: Dict
-    ) -> Dict:
-        """Creates a standardized entry for API components."""
-        return {
-            "type": entry_type,
-            "path": path,
-            "verb": verb,
-            "yaml": yaml.dump(yaml_content, sort_keys=False),
-        }
 
     @staticmethod
     def _normalize_path(path: str) -> str:
@@ -76,4 +71,12 @@ class APIDefinitionSplitter:
 
         if start_index < len(parts):
             return "/" + "/".join(parts[start_index:])
+        return path
+
+    @staticmethod
+    def _get_root_path(path: str) -> str:
+        """Gets the root path from a full path."""
+        match = re.match(r"(/[^/?]+)", path)
+        if match:
+            return match.group(1)
         return path
